split mainwindow across multiple files
