+++ ViAn/GUI/analysiscomponent.h
@@ -1,0 +1,10 @@
+#ifndef ANALYSISCOMPONENT_H
+#define ANALYSISCOMPONENT_H
+#include "mainwindow.h"
+class AnalysisComponent
+{
+public:
+    AnalysisComponent();
+};
+
+#endif // ANALYSISCOMPONENT_H
