+++ ViAn/GUI/documentlistcomponent.h
@@ -1,0 +1,10 @@
+#ifndef DOCUMENTLISTCOMPONENT_H
+#define DOCUMENTLISTCOMPONENT_H
+#include "mainwindow.h"
+class DocumentListComponent
+{
+public:
+    DocumentListComponent();
+};
+
+#endif // DOCUMENTLISTCOMPONENT_H
