+++ ViAn/GUI/drawingcomponent.h
@@ -1,0 +1,13 @@
+#ifndef DRAWINGCOMPONENT_H
+#define DRAWINGCOMPONENT_H
+#include "mainwindow.h"
+#include <QColor>
+#include <QMainWindow>
+#include <QColorDialog>
+class DrawingComponent
+{
+public:
+    DrawingComponent();
+};
+
+#endif // DRAWINGCOMPONENT_H
