+++ ViAn/GUI/filehandlercomponent.h
@@ -1,0 +1,10 @@
+#ifndef FILEHANDLERCOMPONENT_H
+#define FILEHANDLERCOMPONENT_H
+#include "mainwindow.h"
+class FilehandlerComponent
+{
+public:
+    FilehandlerComponent();
+};
+
+#endif // FILEHANDLERCOMPONENT_H
