+++ ViAn/GUI/projectcomponent.h
@@ -1,0 +1,13 @@
+#ifndef PROJECTCOMPONENT_H
+#define PROJECTCOMPONENT_H
+#include "qtreeprojectitem.h"
+#include "makeproject.h"
+class ProjectComponent
+{
+    QTreeProjectItem project_tree;
+public:
+    ProjectComponent();
+
+};
+
+#endif // PROJECTCOMPONENT_H
