+++ ViAn/GUI/qtreeprojectitem.cpp
@@ -1,0 +1,5 @@
+#include "qtreeprojectitem.h"
+
+QTreeProjectItem::QTreeProjectItem()
+{
+}
