+++ ViAn/GUI/qtreeprojectitem.h
@@ -1,0 +1,10 @@
+#ifndef QTREEPROJECTITEM_H
+#define QTREEPROJECTITEM_H
+
+class QTreeProjectItem
+{
+public:
+    QTreeProjectItem();
+};
+
+#endif // QTREEPROJECTITEM_H
