+++ ViAn/GUI/qtreevideoitem.h
@@ -1,0 +1,10 @@
+#ifndef QTREEVIDEOITEM_H
+#define QTREEVIDEOITEM_H
+
+class QTreeVideoItem
+{
+public:
+    QTreeVideoItem();
+};
+
+#endif // QTREEVIDEOITEM_H
