+++ ViAn/GUI/videocomponent.h
@@ -1,0 +1,10 @@
+#ifndef VIDEOCOMPONENT_H
+#define VIDEOCOMPONENT_H
+#include "makeproject.h"
+class VideoComponent
+{
+public:
+    VideoComponent();
+};
+
+#endif // VIDEOCOMPONENT_H
